Add support for non-anonymous supportclients.
Introduced the 'treat as supportclient' permission. If the user has this
permission, their incoming emails are expected to be support requests and
the 'owner-email' attribute is filled in the issue.
The  'treat as supportclient' is added to the 'Anonymous' role by default
(can be removed to avoid receiving support from unknown emails).